requirements.txt
dash
pandas
dash_bootstrap_components
pickle
lightgbm
plotly.express